feat: rename projects WebApi -> Golad and WebApi.Tests -> Golad.Tests
