+++ Golad/ClientApp/home/GameTitleCard.jsx
