+++ Golad/Controllers/SampleDataController.cs
@@ -1,10 +1,9 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 
-namespace goladspa.Controllers
+namespace Golad.Controllers
 {
     [Route("api/[controller]")]
     public class SampleDataController : Controller
