+++ Golad/Program.cs
@@ -1,7 +1,7 @@
 ﻿using System.IO;
 using Microsoft.AspNetCore.Hosting;
 
-namespace WebApi
+namespace Golad
 {
     public class Program
     {
