+++ Golad/Startup.cs
@@ -5,7 +5,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 
-namespace WebApi
+namespace Golad
 {
     public class Startup
     {
