+++ Golad/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "goladspa",
+  "name": "golad",
   "version": "0.0.0",
   "devDependencies": {
     "aspnet-webpack": "^1.0.27",
