+++ Golad/webpack.config.js
