+++ Golad/webpack.config.vendor.js
