+++ golad.sln
@@ -2,9 +2,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.26228.4
 MinimumVisualStudioVersion = 15.0.26124.0
-Project("{13B669BE-BB05-4DDF-9536-439F39A36129}") = "WebApi", "WebApi\WebApi.csproj", "{1A94A1EF-5F51-4F79-AE8D-7923A1B8F7C1}"
+Project("{13B669BE-BB05-4DDF-9536-439F39A36129}") = "Golad", "Golad\Golad.csproj", "{1A94A1EF-5F51-4F79-AE8D-7923A1B8F7C1}"
 EndProject
-Project("{13B669BE-BB05-4DDF-9536-439F39A36129}") = "WebApi.Tests", "WebApi.Tests\WebApi.Tests.csproj", "{817300FA-A2BC-4785-8C79-3FF44A271E08}"
+Project("{13B669BE-BB05-4DDF-9536-439F39A36129}") = "Golad.Tests", "Golad.Tests\Golad.Tests.csproj", "{817300FA-A2BC-4785-8C79-3FF44A271E08}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
